add W0404/W0406 lint rules with file filtering and calibrated thresholds
- W0406 (star_import): detect wildcard imports, skip __init__/mod.spl facades
- W0404 (wide_public): detect over-exported modules, threshold 15→30, skip facades
- Text-scanning versions wired into EasyFix check_all_rules registry
- AST-based versions in 35.semantics/lint/ for compiler integration
- Tests for wide_public (threshold=30 verified)

+++ src/compiler_rust/runtime/src/value/ffi/random.rs
@@ -82,7 +82,7 @@ pub extern "C" fn rt_random_randint(min: i64, max: i64) -> i64 {
     if min > max {
         return min;
     }
-    let range = (max - min + 1) as i64;
+    let range = max - min + 1;
     let rand_val = rt_random_next();
     min + (rand_val % range)
 }
